Referneces added
Though they don't work in the C# file yet. They are just checked and then compiled to binary as something that looks a bit like a list (like most of our data). A few other refactorings and changes too. Util is now in a PyMark folder which makes the main folder a bit cleaner.

Most of the code is also now in the Compiler class in the PyMark folder. PyMark.py is mainly used for command line hooks.